Move utils to detail folder
