+++ include/snowplow/detail/utils/utils_macos.mm
