+++ include/snowplow/detail/utils/utils_macos_interface.h
