+++ include/snowplow/payload/event_payload.cpp
@@ -12,7 +12,7 @@ See the Apache License Version 2.0 for the specific language governing permissio
 */
 
 #include "event_payload.hpp"
-#include "../utils.hpp"
+#include "../detail/utils/utils.hpp"
 
 using namespace snowplow;
 using std::to_string;
