+++ test/payload/event_payload_test.cpp
@@ -12,7 +12,7 @@ See the Apache License Version 2.0 for the specific language governing permissio
 */
 
 #include "../../include/snowplow/payload/event_payload.hpp"
-#include "../../include/snowplow/utils.hpp"
+#include "../../include/snowplow/detail/utils/utils.hpp"
 #include "../catch.hpp"
 #include <map>
 #include <string>
